--- app/components/components/addressComplete.vue
@@ -1,33 +1,0 @@
-<template>
-  <Page class="seccion-register-bg-invert">
-    <label text="jshdk" />
-  </Page>
-</template>
-
-<script>
-import HelpComponent from "~/components/components/HelpComponent.vue";
-import { apiPost,apiGet} from '~/resource/http';
-export default {
-  components: {
-    HelpComponent,
-  },
-  data() {
-    return {
-      
-    };
-  },
-  computed:{
-   
-  },
-  async mounted(){
-   
-  },
-  methods:{
-    
-  }
-};
-</script>
-
-<style scoped>
-
-</style>
